merge software, add protothreads
This commit concludes the basic construction of the new software architecture.

The main file now only contains the setup and loop functions for better overview.
In addition it contains only the global variables that are directly used or might
have to be edited in future versions (e.g. pins and adresses for peripherals).

The other functionalities are split over the files menu, sensors and wifi_communication.
	- menu contains the functions to manage the direct communication with the
	user at the mC (via the LCD and the buttons).
	- sensors contains the functions to extract the data from the sensor readings
	and store them into a fifo storage.
	- wifi_communication contains functions to manage the communication of the mC
	with the user via the internet.
Other global variables are directly declared and initialised in the files they are
(mainly) used.

The addition of protothreads allows the program to immitate a system with multiple
cores. Instead of real parallel computation, the Library provides linear code
execution for event driven systems (see official doc).
Instead of blocking the entire mC while a delay is active, only the specified
protothread is blocked while PT_SLEEP is active. Other protothreads and functions
can be called and are not impacted.

System not yet tested/debugged.